Changing logical axis order to C order was fortran
It seems that some numpy operations are not as smart as
to take the lesser stride to change first whatever
the axis order. For example, when dealing with temporaries
or array indices they are created using the normal sequence.
So using always the default numpy sequence for logical
indexes makes it quite faster.

This has the drawback humans are used to say x,y not y,x.
No need now to swapaxes for any but for pygame which is done just once.